@@ -14,60 +14,60 @@
   "lint-staged": {
     "*.{js,ts,css,scss,md}": [
       "npx prettier --write",
-      "npx lint --fix"
+      "npx lint"
     ]
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
+    "@angular/animations": "^18.2.13",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
     "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
-    "@fontsource/roboto-mono": "^5.0.16",
-    "@ngneat/helipopper": "^8.0.3",
-    "body-parser": "1.20.3",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
+    "@fontsource/roboto-mono": "^5.2.6",
+    "@ngneat/helipopper": "^11.1.1",
+    "body-parser": "2.2.0",
     "cors": "^2.8.5",
     "masonry-layout": "^4.2.2",
-    "material-icons": "^1.13.12",
+    "material-icons": "^1.13.14",
     "ng-in-viewport": "^16.0.0",
     "ngx-build-plus": "^16.0.0",
     "ngx-masonry": "^14.0.1",
-    "rxjs": "~7.5.0",
-    "tslib": "^2.3.0",
+    "rxjs": "~7.8.2",
+    "tslib": "^2.8.1",
     "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-builders/jest": "^16.0.0",
-    "@angular-eslint/builder": "16.0.0",
-    "@angular-eslint/eslint-plugin": "16.0.0",
-    "@angular-eslint/eslint-plugin-template": "16.0.0",
-    "@angular-eslint/schematics": "16.0.0",
-    "@angular-eslint/template-parser": "16.0.0",
-    "@angular/cli": "^17.3.17",
-    "@angular/compiler-cli": "^17.3.12",
-    "@angular/localize": "^17.3.12",
-    "@types/jasmine": "~4.3.0",
-    "@types/jest": "^29.5.3",
-    "@typescript-eslint/eslint-plugin": "5.62.0",
-    "@typescript-eslint/parser": "5.62.0",
-    "autoprefixer": "^10.4.20",
-    "eslint": "^8.44.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-prettier": "^5.0.0",
-    "husky": "^8.0.0",
-    "jasmine-core": "~4.5.0",
-    "jest": "^29.6.2",
-    "lint-staged": "^14.0.0",
-    "postcss": "^8.4.47",
-    "prettier": "^3.0.2",
-    "prettier-eslint": "^15.0.1",
-    "tailwindcss": "^3.4.14",
-    "typescript": "~5.4.5"
+    "@angular-builders/jest": "^18.0.0",
+    "@angular-eslint/builder": "18.4.3",
+    "@angular-eslint/eslint-plugin": "18.4.3",
+    "@angular-eslint/eslint-plugin-template": "18.4.3",
+    "@angular-eslint/schematics": "18.4.3",
+    "@angular-eslint/template-parser": "18.4.3",
+    "@angular/cli": "^18.2.20",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/localize": "^18.2.13",
+    "@types/jasmine": "~5.1.8",
+    "@types/jest": "^30.0.0",
+    "@typescript-eslint/eslint-plugin": "8.36.0",
+    "@typescript-eslint/parser": "8.36.0",
+    "autoprefixer": "^10.4.21",
+    "eslint": "^9.30.1",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-plugin-prettier": "^5.5.1",
+    "husky": "^9.1.7",
+    "jasmine-core": "~5.8.0",
+    "jest": "^29.7.0",
+    "lint-staged": "^16.1.2",
+    "postcss": "^8.5.6",
+    "prettier": "^3.6.2",
+    "prettier-eslint": "^16.4.2",
+    "tailwindcss": "^3.4.17",
+    "typescript": "~5.5.4"
   }
 }
